requirements.txt
Django==3.2.9
djangorestframework==3.12.4
django-cors-headers==3.10.0
gunicorn==20.1.0
pytest==7.1.2
pytest-django==4.5.2